fix bug and refactor handle test
