+++ src/app/services/page.tsx
@@ -2,33 +2,17 @@
 
 import Drawer from '@/components/Drawer';
 import Map from '@/components/Map';
-import { useState } from 'react';
-import { dance3, runTest } from '../actions';
-import { act } from 'react-dom/test-utils';
+import { useActionHandler } from '../handleAction';
 
 export default function ServicesPage() {
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [dialogMessage, setDialogMessage] = useState('');
-
-  const handleAction = async (actionType: String) => {
-    try {
-      if ((actionType = 'test')) {
-        await runTest();
-      } else if ((actionType = 'dance3')) {
-        await dance3();
-      }
-      setDialogMessage('Test erfolgreich ausgeführt');
-    } catch (err: any) {
-      setDialogMessage('Fehler: ' + err.message);
-    }
-    setDialogOpen(true);
-  };
+  const { dialogOpen, setDialogOpen, dialogMessage, handleAction } =
+    useActionHandler();
   return (
     <div
       className="flex flex-col"
       style={{ minHeight: 'calc(100vh - 140px)' }}
     >
-      <Drawer handleAction={handleAction} />
+      <Drawer />
       <div className="flex flex-col grow">
         <Map />
         <div className="flex flex-row justify-around m-5">
